Major
I actually think I discovered the issue, entirely by accident!

When I was generating a new World, I deleted the Weeaboo Faction for a test, and then added it back. Suddenly, everything worked.

I think what happened was that the Weeabooes were generating before the Kuge Shogunate was, and thus, because the Kuge Shogunate was not in the game, there were no Royal Titles to give the Weeabooes.

Apparently, the issue was on your end, not mine: the three Factions that you added all have the same line of code.

It's this: "<configurationListOrderPriority MayRequire="heis.edo.themed.expansion">1000</configurationListOrderPriority>".

When I made the Weeabooes' CONFIGURATIONLISTORDERPRIORITY higher than your Mod (I chose nine thousand nine hundred ninety nine, versus your thousand), suddenly everything worked. Apparently, the problem is that your Faction is the last one that "Rimworld" generates, meaning that anything that depends on it cannot actually call it as a Base or, in this case, render its Royal Title tables.